+++ localization/sparse_mapping/CMakeLists.txt
@@ -185,7 +185,7 @@ target_link_libraries(remove_low_movement_images
   sparse_mapping ${catkin_LIBRARIES})
 
 ## Declare a C++ executable: remove_rotation_only_images
-add_executable(remove_rotation_only_images tools/remove_rotation_only_images.cc)
+add_executable(remove_rotation_only_images tools/remove_rotation_only_images.cc tools/utilities.cc)
 add_dependencies(remove_rotation_only_images ${catkin_EXPORTED_TARGETS})
 target_link_libraries(remove_rotation_only_images
   sparse_mapping ${catkin_LIBRARIES})
